Tweak the thresholds for start/end of zero gravity
Previously there was a problem where some throws would measure
as very short (often less than 0.01 meters). I believe this
was caused by ending the 'free-fall' too soon by a spinning phone
or other mis-measurement of the accelerometer. This commit adds
a second threshold so there is one threshold to start the free-fall
and a second, higher one for ending it.